+++ app/account/views.py
@@ -1,13 +1,14 @@
 from flask import Blueprint, request, render_template, flash, redirect, url_for, g
 from werkzeug.exceptions import abort
-from app.auth.controllers import login_required
+from app.auth.views import login_required
 from app import db
  
 # Define the blueprint: 'register', set its url prefix: app.url/register
-mod_return = Blueprint('return', __name__, url_prefix='/return')
+mod_account = Blueprint('account', __name__, url_prefix='/account')
 
 
-@mod_return.route('/')
+@mod_account.route('/')
 @login_required
 def index():
-    return render_template('return/index.html')
+    return render_template('account/index.html')
+
