+++ app/register/views.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, render_template, flash, redirect, url_for, g
 from werkzeug.exceptions import abort
-from app.auth.controllers import login_required
-from app.register.models import Book
+from app.auth.views import login_required
+from app.models import Book
 from app import db
 import requests
