+++ app/search/views.py
@@ -1,14 +1,13 @@
 from flask import Blueprint, request, render_template, flash, redirect, url_for, g
 from werkzeug.exceptions import abort
-from app.auth.controllers import login_required
+from app.auth.views import login_required
 from app import db
  
 # Define the blueprint: 'register', set its url prefix: app.url/register
-mod_account = Blueprint('account', __name__, url_prefix='/account')
+mod_search = Blueprint('search', __name__, url_prefix='/search')
 
 
-@mod_account.route('/')
+@mod_search.route('/')
 @login_required
 def index():
-    return render_template('account/index.html')
-
+    return render_template('search/index.html')
